requirements.txt
django
gunicorn
django-cors-headers
django-safedelete
django_cron
django-filter
djangorestframework